+++ packages/components/src/misc/CardEmptyCover/CardEmptyCover.js
@@ -17,7 +17,7 @@ import { AssetExpressTaskIcon } from '../FileIcon/AssetExpressTaskIcon';
 import { AssetModuleIcon } from '../FileIcon/AssetModuleIcon';
 import { AssetFeedbackIcon } from '../FileIcon/AssetFeedbackIcon';
 import { FileItemDisplay } from '../../informative/FileItemDisplay';
-import { AssetDocumentIcon } from '../FileIcon/AssetDocumentIcon';
+import { AssetDocumentIconSmall } from '../FileIcon/AssetDocumentIcon';
 
 const CardEmptyCover = memo(({ icon, fileType }) => {
   const pairColumnRef = useRef(null);
@@ -28,7 +28,7 @@ const CardEmptyCover = memo(({ icon, fileType }) => {
     { key: 'image', value: <AssetImageIcon height={24} width={24} color={'#878D96'} /> },
     { key: 'bookmark', value: <AssetBookmarkIcon height={24} width={24} color={'#878D96'} /> },
     { key: 'feedback', value: <AssetFeedbackIcon height={24} width={24} color={'#878D96'} /> },
-    { key: 'document', value: <AssetDocumentIcon height={24} width={24} color={'#878D96'} /> },
+    { key: 'document', value: <AssetDocumentIconSmall height={18} width={18} color={'#878D96'} /> },
     { key: 'task', value: <AssetTaskIcon height={24} width={24} color={'#878D96'} /> },
     {
       key: 'taskexpress',