+++ packages/components/src/misc/FileIcon/AssetDocumentIcon/AssetDocumentIcon.constants.js
@@ -10,3 +10,14 @@ export const DOCUMENT_ICON_PROP_TYPES = {
   height: PropTypes.number,
   color: PropTypes.string,
 };
+
+export const DOCUMENT_ICON_SMALL_DEFAULT_PROPS = {
+  width: 18,
+  height: 18,
+  color: '#7E8795',
+};
+export const DOCUMENT_ICON_SMALL_PROP_TYPES = {
+  width: PropTypes.number,
+  height: PropTypes.number,
+  color: PropTypes.string,
+};