+++ packages/components/src/misc/FileIcon/FileIcon.js
@@ -9,7 +9,7 @@ import { FileItemDisplay } from '../../informative/FileItemDisplay';
 import { AssetAudioIcon } from './AssetAudioIcon';
 import { AssetImageIcon } from './AssetImageIcon';
 import { AssetBookmarkIcon } from './AssetBookmarkIcon';
-import { AssetDocumentIcon } from './AssetDocumentIcon';
+import { AssetDocumentIconSmall } from './AssetDocumentIcon';
 
 export const FILE_ICON_DEFAULT_PROPS = {
   size: 16,
@@ -36,7 +36,7 @@ const FileIcon = ({ fileType, fileExtension, label, size, color, iconStyle, ...p
     { key: 'bookmark', value: <AssetBookmarkIcon height={size} width={size} color={color} /> },
     { key: 'path', value: <AssetPathIcon height={size} width={size} /> },
     { key: 'curriculum', value: <PluginCurriculumIcon height={size} width={size} /> },
-    { key: 'document', value: <AssetDocumentIcon height={size} width={size} /> },
+    { key: 'document', value: <AssetDocumentIconSmall height={18} width={18} /> },
     {
       key: 'file',
       value: (